requirements.txt
# Core LLM and LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-experimental>=0.0.45
langchain-openai>=0.1.0
langchain-google-genai>=0.0.6

# Environment variable support (optional for Streamlit Cloud, required for local dev)
python-dotenv>=1.0.0

# Graph visualization


# Web UI
streamlit>=1.32.0
